Adjust tests.
